First steps towards restricting publishing
Policies should not be publishable by the policy writers; only the second set of eyes (who will be a departmental editor) can do that.

Since we don't have any concept of roles yet (although maybe we should add that soon?), this restriction is simply that if you wrote the policy, you cannot publish it.