Clearer flash message when publishing a stale version
When a policy cannot be submitted because it's been changed by someone else, make it clear that we are displaying the *latest* version of the policy.

Because the message needs to communicate something about what the user is seeing, it seems wrong to generate that message in the model. This implementation still feels alright to us.