Rename Policies to Editions.
In preparation for one Policy having multiple editions, we've renamed the Policy model to Edition.  A single Edition has the same behaviour that Policy used to (including the workflow from draft to published).  In a later commit, we intend to introduce a new Policy model that will collect several editions together.

We have renamed Policy to Edition in most places, but left the human readable text the same.  We've also deliberately left the public facing PoliciesController almost unchanged, as different editions are not currently intended to be seen by the public.