Make the fact-check requesting step easier to reuse.
At the moment this means not assuming that `@edition` exists, but assuming that we're at the point in the application where we can immediately request the fact check by entering their email.

I've extracted the steps which set up behind-the-scenes state into a separate step file whose name more clearly communicates something like "an instance of this thing exists, but I make no guarantees about its relationship to anything else". I prefer that to the slightly misleading "I have written a draft policy", even though the factory object has no actually connection to the feature-level author.

I've also reworded the writer step to make it clearer that they are, at that point, "drafting" a policy. I think this helps clarify the point in the flow that the step leaves you, i.e. still on the edit form of a saved draft policy.