Fact checkers should be able to leave comments.
To do this, I've made the FactCheckRequest model a bit more significant; it now stores the email to whom it was sent, and the comments that the recipient leaves.

I had to modify the token generation for FactCheckRequest because base64 would sometimes return "/" characters that broken some of our steps. A simple token seems fine anyway.