DRY up scenarios.
We've chosen to use factories to set-up the world, rather than going through the UI. There are pros & cons either way and we welcome feedback.

The advantage of this approach is mainly performance and perhaps slightly simpler step definitions.

The disadvantage is that since we're not using the UI we're relying on our factories to setup realistic model state.