Allow a writer to attach a file to a policy.
This functionality is very basic at the moment. Notably the files are stored in the filesystem and in the public directory which means they are accessible by the public *before* the policy is published (although there is some security by dint of the obscurity of the URL). While this will almost certainly need to change, the current commit is enough to satisfy the story.