Admins should be able to see when a document has an attachment.
This is important when editing a document, because, unless you can see an existing attachment, you might unintentionally overwrite the existing attachment or submit it to the second pair of eyes without realising it has an attachment.

This is important when viewing a document as a departmental editor, because unless you can see the attachment, you might inadvertently publish a document with an attachment you didn't know about.