Generate document body HTML using Govspeak.
Govspeak includes a few Markdown extensions which we're not sure we need, but it makes sense to use the Govspeak gem rather than the Kramdown gem to keep us more consistent with other GovUK projects and to discourage divergence.

We have left the fact checking comment text unchanged for now.