See published policies in a topic.
After discussion with Pete H, we decided to make the association between topics & documents many-to-many, since we believe some policies will need to appear in multiple policies.

We've changed the implementation of Document.published so it works as a named scope. It's a bit horrible having a SQL query in the method, but anticipate the data model will change and this will hopefully become unnecessary.