Only display each document once on the topic page.
Topic#documents was returning the same document multiple times if it had multiple editions.

For now we've removed the broken Topic#documents association and we're iterating over editions instead of documents. This feels like a further nudge towards the idea that the editions are actually our primary model, but that's for another day.