Associate ministers with an organisation.
A list of ministers is displayed on the organisation page.

There is a possibility we may need this to be a many-to-many association, but I've not yet been convinced this is necessary.

We may also need to mandate that a minister is in an organisation, but that's not obviously true to me either.