A "minister" is really a "role" that a "person" fulfills.
We want to really clarify the relationships between people and organisations/documents in the system. A Minster is really a person fulfilling a particular role (within an organisation). A Person can have multiple Roles, but we believe that a Role belongs to a single Organisation. We've challenged Neil to provide an example where this isn't true.

There may still be space for "Minster" as a higher-level domain object (something encapsulating a particular role and the person currently fulfilling it), but that can be addressed later.