The distinction between people and roles should be clearer in step definitions.
These domain objects are very important, so we can't miss any opportunity to really nail them down in the application. Again, it's possible that we might introduce a higher-level domain object (something like 'Minister') to encapsulate the appointment of a Person to a Role, and therefore avoid having to call `role.person.name`, but that can come later.

It's also worth noting that we hit a Ruby parser issue writing this step; essentially, given this code:

    assert [1,2,3].any? do |x|
      x > 10
    end

the assertion will pass, because the block is actually passed to the `assert` method, and not the `any?` method. The solution is either to assign the result of the `any?` method (with the block) to a local variable, and assert against the local variable, or use braces rather than do/end.