Topic.with_published_documents returns a distinct list of topics.
I changed the test to expose the problem with the existing implementation.  As well as adding a topic with multiple published documents I had to remove our use of #to_set as that was removing duplicates from the array and it was the duplicates that I needed to expose the bug.