Rename Role to MinisterialRole
It became overly irksome to want to talk about "Ministers" in the UI, but deal with "Roles" in the backend. This is a compromise that better models the actual meaning of the Role - we don't currently know of any non-Ministerial roles that people will have which can also be associated with documents.