Remove 'The Rt Hon' from people's names.
The Right Honourable refers to membership of the Privy Council, life membership of which is granted to all senior ministers.  Rather than discarding this information, instead I've chosen to add a privy_councillor column, and populate it both for seed data and existing people.  It seems wasteful to throw away data that may be useful one day in the future.