A simple minister page
This shows the documents (policies, publications) associated with each ministerial role.

Because a Person can have multiple MinisterialRoles, it's possible (and indeed true for Theresa May, for example) that they will appear twice in the /ministers index page. We need to think a bit more about how we want to present "ministers" with multiple roles.