Clarify which actions require a logged in user.
When we modified this test earlier, we were struck that an Admin controller didn't have any authentication around it. Only after looking in more detail was it clear that only the `create` action should require a login; the other actions are authenticated using a token.

We still need to ensure a user is logged in for `create`, but this should hopefully clarify things in the meantime.