Avoid lots of queries to load organisations.
This happens when calling `to_s` on every `MinisterialRole` loaded, as is the case when generating the `select` element in the document edit view. Including `organisations` in the query that's used in this specific case is an unobtrusive solution; in the future we might want to always load organisations with ministerial roles.