Fix the build, and deployments.
Obvious in retrospect, this prevents the app from starting before the corresponding migration was run. We're still not sure about this approach to be honest.

My thanks to Chris for helping bounce around some ideas; if we look back on this in the future and its folly is laid bare, I accept all responsibility.