Limit applicable nations to policies
There's talk about a news story usefully being associated with a nation, but it's not clear that this cannot be inferred from the link between a news story and a policy. Ditto for other document types. If a compelling reason for a first-class association between a nation and other document types, we can always revert this commit.