Render conflicting document without using a form.
This addresses the bug identified by Pete where since the multi select boxes for the conflicting document were disabled, it was impossible to scroll to see which options had been selected. This is achieved by using a separate partial for the conflicting document which is loosely based on the document show page.

Unfortunately this fix results in a page that is even less pretty than it was, but, since the page already had some problems in this regard, we've decided that this can be addressed in separate commits.

Note that we decided to render the original markdown for the conflicting document body to match the value in the textarea input in the form.