Avoid an error when creating new drafts.
When building, we were seeing a "Supporting documents is invalid" error flash. I think this is because the supporting documents were being added to a document that wasn't saved, and so couldn't be saved themselves (they had no `document_id`). Actually creating the new draft and then creating copies of the supporting documents avoids this.