Reduced length of document conflict flash message.
Flash notices should be short bursts of non critical information (Plus they disappear!) ... the length of the previous one was wrapping and looked horrible.

I've replaced it with a short message and added a notice box at the top of the edit page which describes the steps the user needs to take to resolve the conflict.