Naive introduction of ministerial roles.
All tests are passing, but we're not correctly handling the possibility of multiple appointments for a given role or person.

The plan is to introduce appointment started_at and finished_at attributes so we can work out which appointments are current and which are historic.

This was a first step towards modelling minister speeches which should map to an appointment, not just a role or person.