Add related documents to existing documents.
This is possibly too much for now, but I think it is pretty simplistic still, and possibly poorly named (in reference to the document associations).

But it will allow any document to be related to any other document .. I think next I will add some scopes to the various types so they will easier to pull out on the front end.