Show who delivered speech and when & where it was delivered on admin page.
I hope the markup isn't too horrible. My main thought was that this should be part of the document article and not in a separate section, but this may be wrong.

I wonder whether "Written by" is so appropriate for speeches, given that the speech may have been written by someone other than the person entering it into the system. But perhaps this is an issue for all documents...?