Ensure we're using the document title as the basis of the DocumentIdentity slug.
The DocumentIdentity slug was incorrectly being based on an empty string.  This is because we were building the DocumentIdentity at the point the Document was initialized and the DocumentIdentity was then being saved first, i.e. before it knew about its child documents.  We've got around this by building the DocumentIdentity just before the validation of the newly created Document, and by supplying the document's title as the sluggable string.