Use url and filename to generate attachment links.
There seems to be an issue in carrierwave where after reloading the attachment model, the filename is lost.  It is apparently recommended to mount the attachment onto an attribute with a different name than the active record attribute name (hence carrierwave file), but even doing this didn't totally fix the problem, so have had to resort to writing our own filename method.