Fact checkers should be able to comment on supporting documents.
The simplest possible implementation of this is to just show the contents of the supporting documents, and have the fact checkers gather all their comments together. While this is clearly very basic and puts some burden of clarity on the fact checkers, we'd like to get people using this before we prioritise building something more elaborate.