Published documents shouldn't link to the admin interface.
Editors need to link to other documents within the system.  Suzanne has started doing this by copying the URL of an admin preview and using that in the body of another document.  Rather than try to come up with some clever interface to make this easier, and rather than change Suzanne's behaviour, we've chosen to rewrite these links in the public view of these documents.  We've introduced the `govspeak_to_admin_html` helper that renders all the links as they are.  We'll expand upon this in a future commit to highlight those links that would not be rendered in the public view of the document.

This clearly isn't particularly efficient but is probably OK for now.