Highlight the state of linked documents in admin preview.
It's useful for writers to know when a document they are linking to won't be available when this document is published (and so the link would be removed). We also think it's useful to provide them with a link to both the admin preview of the current draft of a document *and* a link to the public version, to give them the best idea of how their document will be linked when it is published.