Don't try to rewrite links that aren't in our admin.
This was causing a bug on staging, as links to external sites were being processed (and of course the "Document" with ID "blah.mp3" was not found).

I think Chris and I originally had restricted the links to the admin interface, but our switch to the Nokogiri approach left that regexp unused.