Present writers with some guidance about GovSpeak
In the future we might want to find a way to keep the help visible while the writer scrolls around, but this will do for now. It's also present at the bottom of the form for clients without javascript.

I had to insert the link into the label in a very hacky way, because I wanted to preserve the default behaviour of inferring the label text from the method, but also add the link into the label. If anyone can think of a better way...