Add links to the public version of the document.
It's useful to be able to get to the public version of the document you're editing/viewing in the admin interface.

*NOTE* We've created a generic 'link_to_public_document' partial that applies to all documents apart from speeches.  Speeches have their own 'link_to_public_document' partial as we have to cast the various speech subtypes as actual Speech objects before linking to their citizen facing public view.

We've played with an approach to testing shared behaviour across all controllers that subclass the documents controller.  We're not entirely convinced that it's the best approach but think it's better than duplicating the tests in each case.