Documents should allow multiple attachments.
Now a document can support multiple attachments. We had to use a `has_many :through` relationship because multiple published versions of a document can relate to the same attachments.

I haven't made any changes to the interface to allow upload of multiple attachments at the same time; at the moment, this can be achieved by uploading an attachment, saving, and then adding the new attachment as before. This can be repeated for as many attachments as required.

In the future we could improve the interface to allow uploading multiple attachments in one go; lets add a different story for that.