Ensure Speech knows about all its subclasses.
ActiveRecord::Base.subclasses depends on each of the subclasses being defined. Currently all the Speech subclasses are defined in app/models/speech.rb, so when we call Speech.subclasses, all the subclasses are defined. We've added this test in case someone moves any of them into another file.

The reason it's important is that it's used in the select input field for creating and editing a speech.