Allow force publishing of unsubmitted documents, and documents where the author is also the current user.
Rather than using the javascript confirmation box (which is kind of clunky), I tried very hard to add an intermediate publication confirmation page, with a warning.  However, getting the different document types to render correctly proved to be a nightmare due to partial paths and the like.  I might take another look if I feel inspiration, but the confirmation box at least provides the functionality.