Take non-logged in users to their destination once logged in.
This functionality was driven by the recently added email notification for fact check responses. Since the notification email contains a link to the admin document page, a user who was not logged in would have been prompted to login and then would have to have clicked on the link again.