Allow writers to reject submitted documents.
An editor can reject a document and give reasons for their rejection.  The original author can then make amendments based on the feedback and resubmit.  When the editor's happy with the document they can publish as normal.

*NOTE* It's presently possible for both writers and editors to reject documents.  We might want to consider restricting this functionality to editors alone.

There's a bit of an implicit link between editorial remarks and the rejected state of a document at the moment.  Specifically, in the '_rejected_by' partial, I use the author of the most recent editorial remark as the person responsible for rejecting the document.  I can envisage us making the editorial remarks functionality more generic at which time we'll need to be careful to unpick this implicit link.

I've assumed that all document types can be rejected, given that they all go through the 'second pair of eyes' stage.

Possible future improvements:

* Use something other than 'rejected' - it's a bit negative.
* Pretty up the display of editorial comments.
* Pretty up the editorial remarks form.  Maybe bring it onto the document page with some Javascript goodness.