DRY up document state change tests.
I think this makes it easier to determine which state transitions are valid.  I noticed that it wasn't possible to delete a rejected document, even though the 'deletable?' method suggested that it was, so I've updated the state transition to make it match the query method.

I've also removed an 'a' from a couple of test names to make them read better.