Editors can order policies.
This is (as usual) the simplest implementation. I've given `document_topics` an ID so we can update them using `nested_attributes`, and the topic form now manages setting the ordering. I've added tests at multiple levels to ensure everything is plugged together properly.

Adding the ordering in the controller feels wrong; it would be simpler to make it the default.

I'd also like to add javascript enhancement to this which would enable drag and drop ordering, but we'll see.