Re-instate published_policies and published_news_articles relationships on Topic
I tried an alternative approach, using a different `PublicTopic` model class with `policies` and `news_articles` methods, but
although it looked promising, I felt it was a little bit too different to the way we've done the rest of the site.  I'm still
interested in the approach though.

At least this will keep James happy ;)