Re-introduced the news articles index.
This was removed in commit 9bf33ca9f687a4410dd4cb37ed9b4801968e152f.  We've decided to reinstate it because we want hackable URLs.  We want a user to be able to strip away the news slug (e.g. /news/news-slug) and end up at a list of news articles (i.e. /news).  We're not making this part of the primary navigation.