DRY up the news articles section.
I'd previously extracted the news articles list and duplicated the news articles sections that contained that list.  The section in the documents/news_articles partial was only displayed if there were news articles, whereas the sections in the announcements index and news_articles index were shown even when there weren't any news articles.

I had originally thought that it might be good to display a message similar to, 'there aren't news articles', in these empty sections but I think for now I'll stick with the convention of not displaying empty sections on the public site.