Refactor grid system into mixins.
So they can be easily used within different CSS selector scopes.

i.e.

  ul.smiles li {
    @include single_column_grid;
  }

so instead of this markup

  <ul class="smiles">
    <li class="g1"></li>
  </ul>

we can have

  <ul class="smiles">
    <li></li>
  </ul>

We are trying to limit the use of the grid classes (g1,g2,g3) to <div> elements which have no purpose but layout/structure, and if we want the same behaviour on a specific element we should define it in it's style instead of plonking a grid class on it which doesn't really add much semantic value.

Also because the grid is based on an elastic folding grid, we need to be careful with adding padding etc the box model of the element which could screw up the elastic width calculations, therefore I'm trying to discourage the use of the grid classes being used on random elements to try and mimic layout choices I have made previously... (jase)