Separate grouped/non-grouped lists.
The list partials (countries, ministerial roles, organisations and topics) were being used for both index pages and in the sidebars of other pages.  We want them to appear grouped on the index pages but not in sidebars so we've split them into two partials (grouped_list and list).

We've got a special case with the ministerial roles, in that we want to be able to customise the content of the li.  We've catered for this by introducing the render_list_of_ministerial_roles helper.