Revert column style for .secondary sections.
Our style changes were too wide reaching in the last commit, so we have reverted the secondary styles to follow their markup classes for grid alignment, and wrapped each of the related document sections on the policy page in g1 divs as an intermediary step.

Ultimately both of the styles that are in effect for 'secondary sections' will be replaced by something more specific as the design evolves.