Extract more test duplication into custom assertions.
Note that in many of these custom assertions, where an option key is provided with a nil value, this is because I want to check an element is present, but I don't care about the value e.g. :-

    assert_select_role_appointment(:new, index: 0, person: nil, started_at: nil)

This means that I want to check that the person & started_at fields are present, but I don't care what value they have.