Only include current appointments in minister select box.
An exception was happening when a non-current ministerial role existed. This was because outside the loop the query was including non-current roles, but inside the loop Role#current_role_appointment only returns something for current roles. Thus for non-current roles, we ended up calling #id on nil and an exception occurred.

By adding an inner join to current_role_appointments outside the loop, we only pass roles with current appointments to the inside of the loop and all is well again.