Changes needed to make admin styles work.
Because #content is used when going through slimmer for the website, a lot of styles were scoped to this to win in the specificity wars.

Admin doesn't need this constraint because it isn't using slimmer, so I've extracted shared css into mixins so they can be included at the necessary specificity as needed in the local scss files.