Show recently published documents relating to policies in a topic.
* Only include documents relating to *published* policies in a topic.
* Only include *published* documents relating to policies in a topic.
* Sort by publication date in reverse chronological order.

There are a few things I don't like about this commit, notably :-

* The introduction of Document::RelatedDocuments#published_documents_related_with and #published_documents_related_to which duplicate code in Document::RelatedDocuments#documents_related_with and #documents_related_to, and spread the concept of "published" around the code base.
* Using Ruby to uniq & sort the result set.

However, this is the best compromise I can come up with that doesn't result in an N+1 query situation and isn't totally incomprehensible.

I think the major thing we could do to simplify the code would be to consider whether we really need the both directions (to/with) of the document relation associations. It looks to me as if we aren't currently using the Document::RelatedDocuments#documents_related_with association, or at least there is no way to use it via the user interface.