Add a way to determine the current Whitehall 'platform'
It can be useful to know the current platform for some configuration tasks.  For example, if storing files on s3, using the platform name as part of the S3 bucket name ensures no conflicts between platforms.  Normally we'd use the rails environment for this purpose, but unfortunately the preview and production platforms both run against the production environment, so a further distinction is needed.