Get our jQuery plugins loading again.
We were seeing Javascript errors stating that methods like `enableGovspeakHelp` weren't available when the document was loaded. It turns out this is to do with closure scoping in Javascript in conjunction with JQuery being somehow put into some kind of no-conflict more. This is James A's guess, but for the sake of actually getting some work done we're not going to explore further.