Temporary script to update organisation_type_id.
We're validating the presence of the organisation_type_id foreign key but not all organisations belong to a type in production.  This currently causes the Document#create_draft method to fail.

We're assigning all organisations to be 'Ministerial departments' as, according to Pete, only three from our current set are different.