Show all speech subtypes as Speech.
It looks a bit odd to include the speech type in the list of recently updated documents, e.g. 'Speech/speaking note published...'.  Given that, when we have icons for the document types we'll probably use the same icon for all speeches (irrespective of type) I think it's best to generalise these to 'Speech'.

Making this commit made me question whether we should be implementing speech types as subclasses rather than an association (e.g. `speech belongs_to: :speech_type`).  I'm not going to change anything just yet but it might be worth thinking about.