Don't error when creating an edition where a submitted one exists.
Instead of assuming that we can jump to a draft if creating a new edition fails, find the actual unpublished edition (which may be submitted or rejected, not just a draft) and redirect to that.

I've had to make DocumentIdentityTest a bit more realistic, since this more general validation no longer allows creation of non-published editions where other ones exist. This is probably for the best.