Prevent errors when rewriting links with smart quotes.
We traced this to an issue with libxml2 < 2.7.8. We investigated updating libxml2 on the servers, but Gareth was understandably cautious about introducing a new version of such a fundamental library.

In the end we discovered the issue could be avoided by passing something other than a String to #replace. We've added a somewhat brittle test that should hopefully explain why it's necessary and prevent regression via refactoring.