Move PublicationMetadatum attributes onto Publication.
* Putting these attributes in their own table was premature optimization
and does not merit the extra complexity introduced by straying from the
Rails way.
* If it turns out we need to be concerned about having large numbers of
documents where many of the attributes are specific to only one type of
document, we might do better to consider a document database.
* I'm keeping almost all the tests unchanged in this commit to
demonstrate that removing this code has not changed the behaviour of the
system. The only test that changed was one that relied on the underlying
implementation - this has been replaced by extending an existing test to
verify the behaviour independent of the implementation.