Simplify creation of documents with nation inapplicabilities.
* Previously we were creating the parent document alone (by calling save
on the document) i.e. without any of its children - we think it was
something to do with the child objects needing the parent object ID,
but we can't remember exactly. And then creating the child objects by
calling update_attributes on the now existing document.
* We've re-written the controller create method in a more Rails-y way
specifically taking advantage of the way the
accepts_nested_attributes_for works. There already is fairly
comprehensive test coverage for this functionality, especially for the
nation inapplicability fields, so we're pretty confident this simplified
version retains the same behaviour.