More Rails-y way of building document attachments.
* Use accepts_nested_attributes_for :attachments so that we will be able
to use the Attachment#file_cache method provided by carrierwave to
recover an uploaded file after validation failure. We're not actually
making that change in this commit, but this commit should make it easier
to add later.
* This also means we can remove the Attachable#attach_file method which
was a bit non-Rails-y.
* We've had to introduce the build_document_attachment method on the
shared NationAppicability controller module which isn't ideal, because
we'd rather this module didn't need to know about attachments.
* Added test coverage for the relevant scenarios for the
Admin::PublicationsController & Admin::ConsultationsController.