Static assets must be served over a url now.
Previously we could develop against static locally by exploting the fact that the asset path could also be local to the app. However, now that slimmer actually loads things like the template, we can't do that.

Instead, we have to run static and point our app at it. I've reused the same environment variable for simplicity.

So, to develop against a local version of static:

    $ cd static
    $ rackup # this starts a server on port 9292 by default

then

    $ cd whitehall
    % STATIC_DEV=http://localhost:9292 script/rails server