Remember uploaded attachment file after validation error.
* Steps to reproduce :-
1. Create or edit a document and choose a file to attach
2. Ensure at least one of the other fields will cause a validation
error, e.g. leave title blank
3. Save the document
4. See the validation error
5. Fix the validation error e.g. fill in the title field
6. Save the document
7. You should see the file attachment has been addded

* Adding the file_cache hidden field means that through the magic of
carrierwave the file uploaded before the validation error is remembered
and is saved when the validation error is fixed. Note that at the moment
there is no visible representation of the chosen file, but that's a
problem for another day.
* The addition of the above hidden field introduced two other problems.
The first was that attempting to remove an attached file caused a
validation error for Attachment#file not being present. We solved this
by adding the `reject_if: :all_blank` option on the
accepts_nested_attributes_for.
* The other problem was we lost the content type when using the cached
version of the attached file after a validation failure. We remedied
this by using the carrierwave `retrieve_from_cache` callback to set the
content type.