Route everything under /government
We had assumed that the router would be able to mount our application under a prefix (i.e. /government) and translate urls so that our app would not be aware of that prefix. Turns out this isn't how the router works; we get the full url as requested by the browser, including the prefix.

This changeset moves every application url under the /government prefix, including asset serving.