Fix assets in production.
It turns out that the sprockets environment is initialised before any of the files in `config/initializers` are loaded, which meant that while asset paths included the router prefix as the site was running, this prefix wasn't included when sprockets was trying to load the digest YAML file (see `actionpack/sprockets/lib/railtie.rb:36).

The simplest way to control the order of initializers is to use an `Engine`, so here we are.