Remove the RouterPrefixEngine.
The presence of this engine was screwing with our rake tasks.  Specifically, it resulted in our default rake task running rake:cucumber twice (i.e the unit, functional and integration tests were ignored).  I'm still to determine exactly why the presence of the engine caused this but did investigate in a vanilla rails app and saw the same wonky behaviour.

The downside of this commit is that the code dealing with the presence of the router is split across two files (`application.rb` and `router_prefix.rb`) but I think the benefit of working rake tasks outweighs that pain.