DRY up 'formatting help'.
It didn't feel right that you had to pass both the :help flag and add the 'govspeak' class to a text_area in order to have the formatting help displayed.  I've changed it so that you only have to pass the :help flag.

I'm using `#stringify_keys!` in `#text_area` to avoid problems with options being passed as strings and symbols.