Allow multiple 'formatting help' sections on one page.
The previous implementation of the dynamic govspeak helper would move the single copy of the help content (identified by '#govspeak_help') to just below the 'formatting help' link.  This causes a problem with pages containing multiple help links as the single copy of the help content will be moved below the *last* 'formatting help' link.

I've fixed this by taking a copy (using `clone()`) of the help content for each of the 'formatting help' links on the page.  I've added a test for this behaviour and re-jigged the test suite to better support the writing of this new test.