Allow policy areas to relate to each other.
* This is so we can display a list of links to related policy areas from a
policy area page.
* Rather than follow the approach taken by related documents, whenever a
relationship from policy area A to policy area B is added, we
automatically add the inverse relationship from the policy area B to
policy area A. This means the relationship is always effectively
bi-directional and it doesn't matter which way round it is added.
* This should avoid the problem where having related policy area A to
policy area B and then editing policy area B, you don't see policy area
A in the list of related policy areas.
* It also means we only need one has_many and one has_many :through
association rather than two of each. And there is no need for a method
to combine the results from the two has_many :through associations into
one array in memory discarding duplicates. All-in-all it feels like a
simpler implementation.
* The PolicyAreaRelation join model is made read-only once it has been
created so although it can be destroyed, it cannot be modified. This is
so the after_create and after_destroy callbacks can maintain the inverse
relationships consistently.