Clarify relationship between policies and policy areas.
Although we can only associate policies to policy areas, the name of the join table and association still hints at the fact that it used to be possible to associate multiple document types to a policy area.

I plan to change the name of the join table but this first step, of renaming the document association to policy, makes things a little clearer already.