Ensure destruction of inverse policy area relations.
We're using the assignment of related_policy_area_ids in the
Admin::PolicyAreaController to add and remove
PolicyArea#related_policy_areas. The after_destroy callback on
PolicyAreaRelation wasn't being fired when relations were removed this
way, so we've had to introduce a before_remove callback on the
PolicyArea#related_policy_areas association. A script is provided to
remove any relations which may have been left behind when its inverse
was removed before this fix was made.