Show when a policy was updated.
On both the policy itself, and the policy area page.

Strictly speaking the story prescribed adding this information into the 'Recently changed box', but I'm not convinced that makes a great deal of sense; that box contains information about related documents. I have assumed that the real goal is for a user to be able to tell how 'fresh' a policy is, and this is the simplest way of achieving that without blurring the responsibility of other parts of the page.

The display of featured policies on the policy area page needs some attention, but I know that Jason is already working on a different version of the "featured" style, so I think we can wait for that.