A more Rails-y way of making Document#related_documents bi-directional.
* As with PolicyAreaRelations, we ensure that an inverse relation is
created when a DocumentRelation is created and the inverse relation is
destroyed when a DocumentRelation is destroyed.
* This means we can avoid having to combine the results from two
association queries into a single array which meant we were unable to
e.g. use standard Rails scopes on the result.
* It also means that when you associated document A with document B, you
can edit document B and see and/or remove the association with document
A. This was not possible previously and was very confusing.
* To limit the size of this commit, I've aliased some of the old-style
methods to the new methods. This has meant I haven't had to change any
e.g. controller code or controller test code. I plan to fix this in
later commits.
* Since we've added validation to DocumentRelation, we need to ensure
all existing data is valid. I've provided a script
ensure_validity_of_document_relations.rb which I've already run
locally against preview and production data and no invalid records
were found. I plan to run this script again on deployment to
double-check.
* Since we now ensure inverse relations are created for all
DocumentRelations, we need to create inverse relations for all existing
relations. I've provided a script create_inverse_document_relations.rb
which I plan to run on deployment.